refactor(ci): Add retries to steps that could fail due to network hiccups (#22207)
## Description

Adds `retryCountOnTaskFailure` to several tasks in ADO pipelines (mostly
related to the performance benchmarks pipeline) that could fail due to
network hiccups and are safe to retry. This is a pattern we already use,
mostly for `npm i`.

This was motivated by a recent transient failure (ECONNRESET during an
npm install) in the performance benchmarks pipeline.