refactor: Address CredScan warning in server pipelines (#22179)
## Description

This PR fixes the CredScan warnings we were getting in the server
pipelines, before they become a blocker that makes the pipeline runs
fail.

The auto-injected CredScan task in server pipelines was complaining
about things that we had already indicated should be skipped (through
the CredScanSuppressions.json file). Turns out that for docker builds,
the file is expected in the "root context" for the docker build, not at
the root of the repo like it is for some other auto-injected tasks. This
PR makes it so we copy the file to the necessary new location in the
server pipelines.

It also replaces a bunch of fake usernames/passwords in a file's
comments with "PLACEHOLDER" which the CredScan task automatically skips
(pro-tip: don't use "PLACEHOLDER" as your actual password 😄).

Finally, it adds more suppressions for files that are part of test code
in some server dependencies.