refactor(ci): Update credscan suppressions to handle historian and gitrest (#22211)
## Description

Follow up to https://github.com/microsoft/FluidFramework/pull/22179,
addressing the same warning for the historian and gitrest pipelines
(gitssh didn't even have warnings).

The same test files for the same dependencies show up in the builds for
routerlicious (server), historian, and gitrest, and the credscan file
was hardcoding the full path for the routerlicious case. Using path
"suffixes" is supported, so now those files are correctly skipped for
all three pipelines.