tree: add and use create utils (#22188)
## Description

Add some simple tree inspired tree creation utilities.

This focuses on two formats:

1. "insertable" which can be json compatible if you special case handles
and works well for unambiguous schema. Similar to using schema
constructors, but can construct cursors, as well as deal with unions and
undefined at the root.
2. "VerboseTree": handles ambiguous schema, and is a simple self
describing format which is good for persisting, even without schema.
Think of it as JsonableTree for simple-tree.

Not included are APIs to clone trees into these formats for export: such
APIs are planned for the future, and will allow round tripping data.