Prevent moist insertion (#22219)
"Moist" trees - trees composed of both hydrated _and_ unhydrated nodes -
are currently impossible to construct, but the degrees of freedom
permitted when inserting nodes into a SharedTree can imply otherwise to
a user.

For example, it is currently possible to construct an unhydrated tree,
then insert some (non-root) subtree of that tree into the document. This
results in the _subtree_ becoming hydrated, but not the root tree. If
the user retains an existing reference to the subtree then that
reference will now be hydrated, but walking down from the
still-unhydrated root tree will yield a new, different unhydrated
subtree. This is confusing and undocumented behavior - see the new unit
test for a concrete example.

This also:
* Updates one of our asserts to a user error, as it can be easily caused
by user error
* Updates the recursive MapTree walk during proxy binding to be
iterative
* Various code cleanup