refactor(api-markdown-documenter): Add first-step functionality for API Model TSDoc linting (#22150)
Adds function `lintApiModel`, which performs API Model "linting"
validation. For now, it strictly evaluates `{@inheritDoc}` tags to
ensure the target reference is valid within the Model. More features
will be added in subsequent PRs.

Note that the API is not yet-package exported. That will come in a
future PR after the core functionality has landed.

Also updates the `simple-suite-test` test data to include an example of
an `{@inheritDoc}` comment with an invalid target.