fix(build-tools): Filter out empty responses from git ls-files (#22247)
#22226 introduced a bug that caused directory paths to get into the list
of cached files in the Biome task. The root cause was incomplete
handling of the response from `git ls-files`. Empty strings made it into
the list of files, which later resolved as a path to the working
directory - the root of the repo.

The fix is to remove empty strings from the results array.