[main -> 2.2] ci: Bump marocchino/sticky-pull-request-comment action to latest version (2.9.0) (#22262)
Cherry-pick of #22259 and #22264 to the 2.2 release branch.

Updates the marocchino/sticky-pull-request-comment action to its latest
version, 2.9.0. This upgrade was needed because some options used in
recent workflows are not available in the older version.

I confirmed that the commit 331f8f5b4215f0445d3c07b4967662a32a2d3e31
corresponds to the v2.9.0 tag in the
marocchino/sticky-pull-request-comment repo:

https://github.com/marocchino/sticky-pull-request-comment/releases/tag/v2.9.0

I also updated the triggering event to pull_request_trigger from
pull_request.