tree: De-register 'afterSchemaChange' handler when disposing simple tree view (#22276)
## Bug
After a `SchematizingSimpleTreeView` is explicitly disposed, upgrading
its schema via a new view results in double disposal of the original
view.

**Repro:**
1. Create a view via `viewWith`.
2. Dispose the view.
3. Create a second view with the same checkout and upgrade the schema
via `upgradeSchema()`.

## Root-cause
When a `SchematizingSimpleTreeView` creates a `CheckoutFlexTreeView`, it
registers an `"afterSchemaChange"` event handler which disposes the flex
view. However, when the `SchematizingSimpleTreeView` is explicitly
disposed, the event handler is not de-registered. So, when the schema is
upgraded, this event handler is called resulting in disposing it again.

## Fix
Moved the registration of `"afterSchemaChange"` event handlers from
`requireSchema` into the callsite `SchematizingSimpleTreeView::update`
where it is set up to be de-registered when the view is disposed.

Added a test to validate this scenario.



[AB#8916](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/8916)