Added correlationId source tracking (#22280)
## Description

This PR adds telemetry to track the origin of the correlation associated
with an API call by adding a new telemetry prop - `correlationIdSource`.
If the client sends a correlationId in the `x-correlation-id` header or
in the `x-telemetry-header`, then the source is set as
`"correlationIdSource": "client"`. Else the correlationId is generated
by the server and the prop is set as `"correlationIdSource": "server"`.

## Breaking Changes

Updates `ITelemetryContextProperties` to include the
`correlationIdSource` property.