build(client): Update build-tools dependencies to latest (0.44) (#22297)
Updated the following:

  client (release group)

Dependencies on build-tools updated:

  @fluid-tools/build-cli: 0.44.0
  @fluidframework/build-tools: 0.44.0
  @fluidframework/bundle-size-tools: 0.44.0
  @fluid-tools/version-tools: 0.44.0

I also made two small changes to the root Biome config and the
test-snapshots Biome config. The configs exposed a case where `git
ls-files` will return a path to a directory - if that directory is a
submodule. These config changes enable the Biome task caching to work
even with this ls-files oddity by ensuring the directory path doesn't
make it into the cached file list.