(fix) getSingleUseLegacyLogCallback would never log in practice (#22303)
The function `getSingleUseLegacyLogCallback` clears itself the first
time it's called - but it should actually clear itself the first time it
actually logs, otherwise it will likely never log (only if the very
first time through happened to be this case)

(Note that this function is still global, not tied to ContainerRuntime
instance. In application usage, this is fine. In tests, it's potentially
problematic, but only if multiple tests are exercising the legacy
codepath, which shouldn't be the case.)