Unit tests for batch validation in container runtime (#22299)
Added unit tests in container runtime that tests that batch validation
works correctly. Most of this validation is in the `ScheduleManager` and
`ScheduleManagerCore` and a couple are in `ContainerRuntime::process`.
Some of these scenarios may be tested in other unit tests or end-to-end
tests but these tests explictly do these validations for the following
scenarios:
- A batch from a client is interleaved from another client's op.
- A batch from a client contains non-runtime ops along with runtime op.
- A batch from a client contains unknonw runtime ops along with known
runtime op.
- A message has batch end metadata without batch begin.

These tests will help with the work of removing `ScheduleManager` and
`ScheduleManagerCore`. The only part of these two classes that is still
needed is batch validation. This work is tracked by
[AB#13509](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/13509).