tree: Remove "any" from tree stored schema system (#22296)
## Description

Remove support for "any" in the flex tree and stored schema layers. This
aligns them with the public simple-tree layer which never supported
"any" (and chant since it requires schema to be reached by walking the
schema from the root).

This should be non-breaking:

If a document existed using "any" in its schema, it would be impossible
to view that document before this change since no simple-tree schema is
compatible with any.

With this change, such a document would become out of schema, failing on
validation, or when its schema is parsed if validation is disabled. This
is a different failure mode for an unusable document, and such a change
should be acceptable especially given such documents should not exist
(as there has been no way to create them since well before FF 2.0)