Signal Telemetry Update + Tests (#22033)
## Description
This PR updates signal telemetry to track more signals and monitor for
abnormal signal behavior. This will allow us to catch more lost signals
and out of order signals.

We look at all incoming signals to see if they are coming in sequential
order. If a signal sequence number(s) is skipped, we log a 'SignalLost'
event and continue to monitor for both new signals and out of sequence
signals. Once we receive an out of order signal (previously logged as
lost), we log a 'SignalOutOfOrder' event. We continue to log a
'SignalLatency' event that details the roundtrip time of a traced
signal, along with the number of signals sent, signals lost, and out of
order signals received the last latency event was logged. We maintain a
lower bound signal monitoring window of `minimumSignalSequenceNumber` --
any signals below this is ignored. This window helps with
disconnect/reconnect scenarios and keeping track of the amount of
signals sent in each 'SignalLatency' period.

Follow up: https://github.com/microsoft/FluidFramework/pull/22023