improvement(generate:typetests): Use only the previous package to generate tests (#22312)
Updates the type test generator to only consider the previous version of
the package when generating tests. This simplifies the code overall and
makes the build process much simpler, since it makes type test
generation dependency-free. All types are pulled from the previous
version of the package, and they're all assumed to be present in the
current version.

The old code made special "removed" test cases when types were not found
in the old package. Those special cases no longer exist. All types in
the previous package are assumed to exist in the current version. When a
type is removed in the current version, then one of the test cases will
fail, and can be suppressed.


[AB#13887](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/13887)