ci: Use outFile flag instead of stdout redirection in flub list (#22319)
`flub list` recently started outputting a new warning, which was
expected, but broke the publish pipeline because it was using stdout
stream redirection to write to the file instead of the `--outFile` flag.
That meant that the warning text made it into the package lists,
breaking the downstream steps. Now it uses the flag, so the resulting
file won't have the warning.