ContainerRuntime: Add loadContainerRuntime free function, and move usages to IContainerRuntime (#22230)
The ContainerRuntime class is very large and complicated while also
being over-exposed, and overused. This reduces our ability to improve
the ContainerRuntime class to increase testability, maintainability, and
extendibility.

This change is the first step on a path to reduce and eventually remove
the exposure of the ContainerRuntime class which will greatly increase
our ability to make changes to its implementation that increase
testability, maintainability, and extendibility.

The core of this change is the introduction of a function that can be
used to replace the existing static load method on the ContainerRuntime
class with a free function which returns an interface, rather than
class:
```
export function loadContainerRuntime(params: {
    context: IContainerContext;
    registryEntries: NamedFluidDataStoreRegistryEntries;
    existing: boolean;
    runtimeOptions?: IContainerRuntimeOptions;
    containerScope?: FluidObject;
    provideEntryPoint: (containerRuntime: IContainerRuntime) => Promise<FluidObject>;
}): Promise<IContainerRuntime>;
```

This load method does not support all known uses of the ContainerRuntime
class and its load method, specifically, scenarios that depend on the
mixin pattern are not handled. However, those usages are less common, so
cases where those features are not needed can move to the new function
and the interface immediately. Once this method exists, we can begin to
ask users of the ContainerRuntime class to move to the new method, and
provide us feedback about obstacles they face, with which we can create
solutions for them that do not depend on the ContainerRuntime class. Due
to this we will explicitly not be deprecating the ContainerRuntime
class, as we know its usage is still needed. Only after we have
confidence that all know scenarios can be covered without using the
class itself should we deprecate, and eventually remove export of the
ContainerRuntime class.

Lastly, this change migrates many of our usages of the ContainerRuntime
class to the new function, and the interface where it is easy and
straightforward to do so. Some remaining usages can be the first set of
scenarios we extend support for. However, there may be cases, like is
test, where we want to use functionality that we never expect to be
public, and that we fully control, these cases may remain, and it will
be our responsibly to maintain them after export of the class is
removed.