(fix) ItExpects exception for OpRoundtripTime error log was using an old eventName (#22353)
The fix: The event name changed at some point but this wasn't updated.

Background: This is a convenience especially when debugging - this event
will fail many itExpects tests if you pause the debugger on a
breakpoint. It only triggers after 5s, no test should be depending on
this behavior (way too slow of a test strategy).