build(client): Remove check:format from lint task (#22355)
Removes the `check:format` task from the `lint` task/phase of the build.
Instead, it's now part of the `build` task.

This has no effect on the local build experience, but in CI, this
effectively means that `check:format` will _only_ run as part of the
separate "checks" stage. It will no longer run as part of the main
build. In fact, we were doing duplicate work, because `check:format` was
running in both the main build and the checks stage.