feat(build-cli): New re-usable flag for parsing version strings (#22360)
Adds a new custom flag, `semverFlag`, which parses the provided value as
a semver. This enables input validation at flag parsing time and ensures
that the command receives a semver.SemVer object instead of just a
string as the version value.

All commands that had a string-type version flag have been updated to
use the new flag.