ci(client): Fix client trigger on any edit to root files (#22203)
Unfortunately the change in #22133 is causing the client build to always
be triggered for PRs, which was not the intended behavior.

Instead of `*`, let's try `*.*`, which I have seen used in other
systems. The docs are not clear at all about the behavior.

I also reordered entries to be alphabetical and made folder entries
consistent.