tree: Destroy newly created nodes when aborting a transaction (#22333)
## Description

Prior to this change, aborting a transaction was implemented by undoing
rather than rolling back the transaction's changes. Because of this, any
nodes created by the transaction would not be destroyed, but put in
detached fields instead. The detached field keys were not tracked by the
DetachedFieldIndex and could end up being reused, which would throw an
exception.