ci: Update workflow permissions (#22373)
One of the output variables had an empty line in it, which meant the
checks weren't getting triggered since the value wasn't the expected
one. Also, the GitHub Actions token doesn't have org:read permissions,
and those can't be given using the `permissions` block in the workflow,
so I created a new narrowly scoped PAT and added it as a secret.