add encoding of identifiers in UniformChunk (#22327)
## Description

This PR allows encoding identifiers into its compressed form in
`UniformChunk`s. The current implementation would also compress strings
which also happens to be compressible ids, but this should not be an
issue, as it is just the in-memory representation, and a check is made
to make sure the values are decompressed before returning.

A follow-up PR will be made to allow a path to opt-in to this encoding.