Update container runtime to only call batch begin and end on schedule manager (#22356)
ScheduleManager is currently called on every op processed by the
container runtime. However, it only cares about the beginning and end of
batches and to do this, it maintains state to identify batch begin and
end.
The container runtime also accumulates batches and only processes it
after the entire batch is available. So, ScheduleManager does not need
to track batches. This PR removes batch tracking from ScheduleManager
and directly calls it when a batch begins and ends.

Also, refactored the processing pipeline in ContainerRuntime::process to
have one `processBatch` function instead of three for runtime,
non-runtime and empty batches. This should make the code cleaner.


[AB#14483](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/14483)