feat(generate:typetests): Add per-package typetest entrypoint config (#22131)
## Background

The type tests are primarily configured in the `typeValidation` field in
package.json. However, the API level setting is only passed via a CLI
flag, so when you make changes to the type tests for a particular
project, you have to update the CLI command. It also means you can't use
the package filtering flags to operate on a subset of packages with the
CLI, which is confusing.

## Changes

The main change is a new typeValidation setting, `entrypoint`, that,
when set, will be used as the "default" entrypoint for that package
(default value is "legacy"). If the CLI `--entrypoint` flag is passed it
is used instead of the default. (The `--level` flag is now a deprecated
alias of the `--entrypoint` flag.)

[AB#7875](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/7875)