Make tests less likely to flake (#22390)
[AB#8942](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/8942)

Prevent tests from flaking.

After taking a look at telemetry for our failing tests for the local
branch in the last 30 days, these were the identified tests. Two of them
were just waiting for the summary op, these stashed ops tests do not
need to use heuristics to create/generate their summaries. We can just
control when those summaries happen as that is what the test seems to
desire.

The other Heuristics based test is testing the summary heuristics, so we
have to manually wait. For that, I've simply just doubled the timeout,
and we should see the summary finish in time.