refactor(build-tools): Reorganize build-tools files in preparation for splitting out fluid-build (#22325)
This change does some internal organization within the build-tools
package to better isolate fluid-build-only functionality from shared
functionality. Everything that is fluid-build only - with the exception
of the config and the "getConfig" function - should be within the
fluidBuild folder and only referenced from other files within that
folder.

This change also:

- Splits the FluidRepo class and the build config into separate files.
- Moves some util functions that are used only within fluid-build or
tasks into the fluid-build-specific taskUtils.
- Removes the unused function `resolveNodeModule` from common/utils.ts

**There is no intended behavior change. It's only meant to reorganize
existing code.**

There are still some circular relationships that need to be
rationalized. For example, the `MonoRepo` and `Package` classes expect
the fluid-build properties to be included in package.json, but that
config should really be specific to fluid-build. That's not addressed in
this change, but it and similar issues will be addressed over time,
starting with the introduction of a shared build infrastructure package
in #22347.