fix(ci): Only try to do code-coverage report steps if the report exists (#22417)
## Description

https://github.com/microsoft/FluidFramework/pull/22346 added steps to
publish the code coverage report in pipeline runs, but we failed to
consider that the builds for some packages never produce code coverage
reports, and those are failing now because we unconditionally try to
find it to publish it. This PR adds a condition to the step that was
failing (a condition several of its "sibling steps" already used, which
guarantees the code coverage report exists) so it only runs when it
should. The condition to publish the artifact was also updated in the
same way.

I also moved the existing `condition` in some steps to before the
`inputs`, where it is more readable (avoids having to "scan back" from
the indented contents of `inputs`, which we usually have as the last bit
for a `task`).