refator: Make e2e tests for OpRoundtripTime telemetry more robust (#22361)
## Description

Refactors the tests for robustness. They are trying to validate that a
`OpRoundtripTime` log is emitted both when a container is created and
when it is loaded. Currently it could happen that if the log was emitted
twice during the creation flow and never by the load flow, the test
would still pass. Now checks are done more explicitly and at better
points during the test execution to prevent this potential false
positive (which, for the record, we don't have evidence ever occurred).

Prompted by [this
comment](https://github.com/microsoft/FluidFramework/pull/19427/files#r1739294835).