fix(client): `exactOptionalPropertyTypes` clean up
Select changes required to support new Presence package being built with `exactOptionalPropertyTypes: true`

When exactOptionalPropertyTypes is false, a property specified as optional (via `?`) may be assigned `undefined`. When true that is not allowed.

Update types to explicitly allow undefined (no effective type change as build continues to use `exactOptionalPropertyTypes: false`) or to properly respect not setting `undefined`. The latter is an implementation change, but conforms to declared specification.