refactor(fluid-build): Mark symlink and depcheck-related APIs deprecated (#22425)
The symlink and depcheck-related functionality in fluid-build is not
maintained and we are moving to rely on the workspace management
features of the package manager instead of doing our own symlink
management. Since these features are not well-maintained, documented, or
used, they are marked as deprecated in this change. We can remove them
in a future change, possibly in 0.46.