fix(generate:typetests): Move type compat exports back to build-tools (#22443)
When the type test code was moved to build-cli, some of the types used
in the generated tests were no longer exported. Because build-cli is an
ESM-only package, and the types need to be used in dual-emit and
CJS-only packages, they have been moved back to build-tools and exported
from there.

Once the new build-infra package in #22347 is merged, these types could
possibly be moved there.