Expose getAliasedDataStoreEntryPoint on IContainerRuntimeBase (#22389)
This change exposes getAliasedDataStoreEntryPoint on
IContainerRuntimeBase, allowing DataStores that only have access to
IContainerRuntimeBase to use it. Previously, this functionality was only
available on IContainerRuntime, which extends IContainerRuntimeBase.

The original placement appears to have been an oversight, as
IContainerRuntimeBase already includes the methods necessary for
creating and aliasing DataStores. It is consistent to allow the creation
and retrieval of aliased DataStores at the same layer.


[AB#14344](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/14344)