fix(generate:releaseNotes): Fix broken TOC links in release notes (#22464)
The release notes TOC links are getting broken when pasted into GitHub
Releases because GitHub prepends all the custom IDs with the string
`user-content-`. With this change, the TOC links will be written with
`user-content-` prepended when the `--headingLinks` flag is provided.

Additional changes:

- Soft (single) line breaks are now stripped so that when rendering in
GitHub Releases, the markdown will render as expected. The markdown
already renders correctly when viewed in the repo release notes files
themselves.
- A "table of contents" link is appended to the bottom of each note.
Clicking it will navigate to the TOC. This makes it easy to jump back
and forth between individual entries.

---------

Co-authored-by: Alex Villarreal <716334+alexvy86@users.noreply.github.com>