tree: Mark internal types as system types (#22470)
## Description

Tree has several types that meet the requirements to be `@system`
(grouped under InternalTypes) but predate the system tag. Now that
system has public facing docs, they should be tagged accordingly.

In the future we may want to remove the InternalTypes grouping/pattern
used on these types as it causes some issues with TypeScript reexports
in customer code, as well as breaks API-Extractor in some cases.
Regardless of if we make a such a change, making them as system better
communicates their status.